mednais-analysis/mednais-sop-marketplace-main/scripts/requirements.txt
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0
emergentintegrations
